Replace bitvector.h/cpp with ptrArgTP type in gc_unwind_x86.h/inl
